+++ pages/info.php
@@ -43,6 +43,7 @@ $content = '
 <li>im Beschreibungsfeld kann mit der Eingabe von ***** (5 Sterne) ein Trenner hinzugefügt werden.</li>
 <li>durch Klick auf "blaue" Aufgaben wird die Beschreibung angezeigt.</li>
 <li>durch Klick auf die Überschriften wird die Tabelle entsprechend sortiert.</li>
+<li>durch Klick auf das Auge (rechts im Tabellenkopf) können die erledigten Aufgaben dauerhaft ein- bzw. ausgeblendet werden.</li>
 <li>Kategorien werden durch den Admin gepflegt.</li>
 
 <hr/>