+++ uninstall.sql
@@ -2,4 +2,5 @@
 DROP TABLE IF EXISTS `%TABLE_PREFIX%aufgaben_aufgaben`;
 DROP TABLE IF EXISTS `%TABLE_PREFIX%aufgaben_kategorien`;
 DROP TABLE IF EXISTS `%TABLE_PREFIX%aufgaben_status`;
+DROP TABLE IF EXISTS `%TABLE_PREFIX%aufgaben_filter`;