+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 Aufgaben - Changelog
 ===
+
+#### 14.01.2016 Version 1.7.1
+
+- Filter Tabelle löschen in uninstall.php
+
+
+#### 14.01.2016 Version 1.7
+
+- Fehler beim Priofilter behoben
+- erledigt Status (Auge) hinzugefügt
+
 #### 14.01.2016 Version 1.5.1
 
 - Fehler gefixt